requirements.txt
certifi==2017.4.17
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.3
django-allauth==0.32.0
django-extensions==1.8.1
django-suit==0.2.25
djangorestframework==3.6.3
idna==2.5
oauthlib==2.0.2
python3-openid==3.1.0
pytz==2017.2
requests==2.18.1
requests-oauthlib==0.8.0
six==1.10.0
urllib3==1.21.1